Workshops/AI/04-Agents-and-RAG/requirements.txt
fastapi
uvicorn[standard]
pydantic
chromadb
sentence-transformers
python-dotenv

# Hugging Face + agent stack
huggingface_hub
langchain
langchain-core
langchain-huggingface
langgraph
sentencepiece 
transformers 
tokenizers
